+++ NEWS.md
@@ -1,3 +1,7 @@
+# cache 0.1.7 
+ 
+ - `use_cache` added ... and now does the right thing for all project types
+
 # cache 0.1.6
 
  - `cache` now returns object invisibly for use in pipes.
@@ -5,4 +9,4 @@
 
 # cache 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+ - Added a `NEWS.md` file to track changes to the package.